requirements-windows.txt
joblib==0.16.0
grpcio==1.32.0
grpcio-tools==1.32.0
matplotlib==3.3.1
pytest==6.2.1
pandas==1.1.0
Pillow==8.1.0
protobuf==3.14.0
tensorflow==2.4.0
numpy==1.19.2
six==1.15.0
scikit_learn==0.23.2
torch==1.11.0
wandb==0.12.11
dvc[ssh]==2.10.0